+++ xsd-examples/cxx/tree/performance/serialization.cxx
@@ -105,7 +105,8 @@ serialization (const char* file, unsigned long iter)
 
     cerr << "  time:           " << time << " sec" << endl;
 
-    double ms (time.sec () * 1000000ULL + time.nsec () / 1000ULL);
+    double ms (static_cast<double> (
+      time.sec () * 1000000ULL + time.nsec () / 1000ULL));
 
     // Calculate throughput in documents/sec.
     //
